Refactor home page to use Suspense for data loading and prefetching at the route loader level for demo purposes
